backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
supabase

langchain
langchain-core
langchain-community
langchain-google-genai

web3